updated profile list view
