+++ Foqos/Components/BlockedProfileView/BlockedProfileRow.swift
@@ -1,53 +1,33 @@
-import FamilyControls
 import SwiftUI
 
 struct ProfileRow: View {
   let profile: BlockedProfiles
+  let isActive: Bool
 
-  var formattedUpdateTime: String {
-    let formatter = RelativeDateTimeFormatter()
-    formatter.unitsStyle = .short
-    return formatter.localizedString(for: profile.updatedAt, relativeTo: Date())
-  }
-
-  var selectedItemsCount: Int {
-    FamilyActivityUtil.countSelectedActivities(profile.selectedActivity)
+  init(profile: BlockedProfiles, isActive: Bool = false) {
+    self.profile = profile
+    self.isActive = isActive
   }
 
   var body: some View {
-    HStack {
-      VStack(alignment: .leading, spacing: 12) {
-        Text(profile.name)
-          .font(.headline)
-
-        HStack(spacing: 4) {
-          Image(systemName: "clock")
-          Text("Updated \(formattedUpdateTime)")
-        }
-        .foregroundStyle(.secondary)
-        .font(.caption)
-      }
-
-      Spacer()
-
-      HStack(spacing: 4) {
-        Image(systemName: "list.bullet.circle.fill")
-        Text("\(selectedItemsCount) items")
-      }
-      .foregroundStyle(.secondary)
-      .font(.subheadline)
-    }
+    ProfileSummaryContent(
+      profile: profile,
+      isActive: isActive,
+      metadata: .appsAndDomains,
+      showsStatusLine: false,
+      layout: .compact
+    )
   }
 }
 
 #Preview {
   let previewProfile = BlockedProfiles(
     name: "⌛ School Hours",
-    selectedActivity: FamilyActivitySelection(),
     createdAt: Date(),
     updatedAt: Date().addingTimeInterval(-3600)
   )
 
   return ProfileRow(profile: previewProfile)
     .padding()
+    .environmentObject(ThemeManager.shared)
 }
